[ext.pages] General fixes and enhancements (#1227)
* adds option to use the navigation button interaction response path to edit the message (used by default), otherwise fall back to message attr edit path

adds `custom_view` parameter to `Page` object, which overrides the custom view for the paginator when that page is being shown

to better accommodate the above item, a new method `update_custom_view` has been added which will change the `custom_view` of a paginator when called.

various other docstring fixes and cleanup

* improved per-page custom views (i.e. make them actually work)

* fix custom views when defined at paginator level

* add behavior of per-page custom views to paginator-level custom view docstring

* add `Page.callback()` method and supporting `Paginator.page_action()` method.

* add `interaction` parameter to `Page.callback()` and `Paginator.page_action()`
add `Paginator.trigger_on_display` (and corresponding `PageGroup`) parameter, which will automatically fire the callback for a `Page` whenever that page is displayed.

* add support for `trigger_on_display` in `Paginator.update()` method

* add fix for rare use case where a paginator had no pages and showed `page_count` as -1

* add default for optional interaction param

* wording change to make it clearer which interaction is being used

Co-authored-by: BobDotCom <71356958+BobDotCom@users.noreply.github.com>