Fixed a bug in ext.pages.Paginator
Before the fix, if use_default_buttons was set to False, then a KeyError would be thrown because page_indicator was not defined in the Paginator.buttons dict since there were no buttons. By editing the if statement before it to check that the default buttons were being used, this error is now solved.