feat: add additional operations to `Permissions`
Implement addition (+), subtraction (-), union (|), intersection (&), and inverse (~) operations on the `Permissions` class and update the documentation accordingly. The addition and union operations are equivalent and their operators are interchangeable (and the latter is consequently omitted from the updated documentation).

This allows for things such as adding two `Permissions` objects together to obtain the combined set of permissions they represent; subtracting two `Permissions` objects to set all permissions that are `True` in one object to `False` in the other; finding the intersection of two `Permissions` objects to obtain the set of permissions that are `True` in both; and inverting a `Permissions` object to set all of its `True` permissions to `False` and vice versa.

The addition, subtraction, union, and intersection operations can operate on a `Permissions` object and a) another `Permissions` object, b) a `flag_value` object, or c) an integer. Reverse methods for each of these operations are also implemented, so both `Permissions.none() + Permissions.view_channel` and `Permissions.view_channel + Permissions.none()` will produce the same result, for example.