Add limit argument to basic autocomplete
The basic autocomplete function removes any values which do not start with the user input string. This makes it impossible to use any other type of filtering method like checking if the user input is contained in an option.

I've added a boolean which allows all values to be used in case a bot developer is filtering their own values. It defaults to True so that the current implementation of the function is matched but can be changed if a developer would like.