refactor!: embed classes (#2063)
Attributes for Author, Footer, Image, Thumbnail, Video, and Provider on the embed will now return `None` instead of `EmbedProxy`, when they are not set. This change is done for correct API reflection.

This allowed the complete removal of `EmbedProxy` in favour of separate classes for the above attributes. This allows for concrete typing support. The new classes are `EmbedAuthor`, `EmbedFooter`, `EmbedMedia` and `EmbedProxy`

`Embed.Empty` and `EmptyEmbed` have now been removed in favour of `None`

Co-Authored-By: Lulalaby <lala@pycord.dev>