fix: Correct regression from pull #2124 (#2126)
Pycord requires typing_extensions as a dependency for python < 3.11,
however not on version >= 3.11. Because the changes in the referenced
pull always tries to load from typing_extensions, they make it so an
exception will be raised if the end user is using python >= 3.11 and
does not otherwise have that library installed when _parse_options is
 run (ie, whenever a command is registered).

This change checks the version of python running Pycord, then imports
stdlib typing if >= 3.11.