[tasks] reset exponential backoff after task runs successfully
The current behaviour of the _loop method (when the reconnect
attribute is set to True) is to have the same ExponentialBackoff
instance for all of its lifetime. A task that has failed n times
over its life-time will thus wait 2**n seconds before retrying, no
matter how many successful instances of happened between the failures.

With this commit, the behaviour is to reset the exponential backoff
after a succesful execution of the task, so that the backoff only
takes place after successive failures.